Use MediaSessionService instead
